requirements.txt
requests>=2.31.0
brotli>=1.1.0
beautifulsoup4>=4.12.0
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.40.0
